+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
   "quickwit-actors",
+  "quickwit-aws",
   "quickwit-cli",
   "quickwit-cluster",
   "quickwit-common",
