+++ quickwit-aws/src/lib.rs
@@ -17,7 +17,5 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-mod api;
-mod helpers;
-pub mod kinesis_source;
-mod shard_consumer;
+pub mod error;
+pub mod retry;
