+++ quickwit-indexing/Cargo.toml
@@ -20,6 +20,7 @@ futures = "0.3"
 itertools = "0.10.3"
 once_cell = "1"
 quickwit-actors = { version = "0.2.1", path = "../quickwit-actors" }
+quickwit-aws = {version = "0.2.1", path = "../quickwit-aws"}
 quickwit-common = { version = "0.2.1", path = "../quickwit-common" }
 quickwit-config = { version = "0.2.1", path = "../quickwit-config" }
 quickwit-directories = { version = "0.2.1", path = "../quickwit-directories"}
