+++ quickwit-indexing/src/source/kinesis/kinesis_source.rs
@@ -25,6 +25,7 @@ use anyhow::Context;
 use async_trait::async_trait;
 use itertools::Itertools;
 use quickwit_actors::{ActorExitStatus, Mailbox};
+use quickwit_aws::retry::RetryParams;
 use quickwit_config::{KinesisSourceParams, RegionOrEndpoint};
 use quickwit_metastore::checkpoint::{CheckpointDelta, PartitionId, Position, SourceCheckpoint};
 use rusoto_core::Region;
@@ -35,7 +36,6 @@ use tokio::time;
 use tracing::{info, warn};
 
 use super::api::list_shards;
-use super::retry::RetryPolicyParams;
 use super::shard_consumer::{ShardConsumer, ShardConsumerHandle, ShardConsumerMessage};
 use crate::models::RawDocBatch;
 use crate::source::{Indexer, Source, SourceContext, TypedSourceFactory};
@@ -94,7 +94,7 @@ pub struct KinesisSource {
     shard_consumers_rx: mpsc::Receiver<ShardConsumerMessage>,
     state: KinesisSourceState,
     shutdown_at_stream_eof: bool,
-    retry_policy: RetryPolicyParams,
+    retry_params: RetryParams,
 }
 
 impl fmt::Debug for KinesisSource {
@@ -120,7 +120,7 @@ impl KinesisSource {
         let kinesis_client = KinesisClient::new(region);
         let (shard_consumers_tx, shard_consumers_rx) = mpsc::channel(1_000);
         let state = KinesisSourceState::default();
-        let retry_policy = RetryPolicyParams::default();
+        let retry_params = RetryParams::default();
         Ok(KinesisSource {
             source_id,
             stream_name,
@@ -130,7 +130,7 @@ impl KinesisSource {
             shard_consumers_rx,
             state,
             shutdown_at_stream_eof,
-            retry_policy,
+            retry_params,
         })
     }
 
@@ -171,7 +171,13 @@ impl KinesisSource {
 #[async_trait]
 impl Source for KinesisSource {
     async fn initialize(&mut self, ctx: &SourceContext) -> Result<(), ActorExitStatus> {
-        let shards = list_shards(&self.kinesis_client, &self.stream_name, None, &self.retry_policy).await?;
+        let shards = list_shards(
+            &self.kinesis_client,
+            &self.stream_name,
+            None,
+            &self.retry_params,
+        )
+        .await?;
         for shard in shards {
             self.spawn_shard_consumer(ctx, shard.shard_id);
         }
