+++ quickwit-indexing/src/source/kinesis/shard_consumer.rs
@@ -25,14 +25,13 @@ use std::time::Duration;
 
 use async_trait::async_trait;
 use quickwit_actors::{Actor, ActorContext, ActorExitStatus, ActorHandle, Handler, Mailbox};
+use quickwit_aws::retry::RetryParams;
 use rusoto_kinesis::{KinesisClient, Record};
 use serde_json::json;
 use tokio::sync::mpsc;
 
 use crate::source::kinesis::api::{get_records, get_shard_iterator};
 use crate::source::SourceContext;
-
-use super::retry::RetryPolicyParams;
 
 #[derive(Debug)]
 pub(super) enum ShardConsumerMessage {
@@ -77,7 +76,7 @@ pub(super) struct ShardConsumer {
     state: ShardConsumerState,
     kinesis_client: KinesisClient,
     sink: mpsc::Sender<ShardConsumerMessage>,
-    retry_policy: RetryPolicyParams,
+    retry_params: RetryParams,
 }
 
 impl fmt::Debug for ShardConsumer {
@@ -107,7 +106,7 @@ impl ShardConsumer {
             shutdown_at_shard_eof,
             kinesis_client,
             sink,
-            retry_policy: RetryPolicyParams::default(),
+            retry_params: RetryParams::default(),
         }
     }
 
@@ -155,7 +154,7 @@ impl Actor for ShardConsumer {
             &self.stream_name,
             &self.shard_id,
             self.from_sequence_number_exclusive.clone(),
-            &self.retry_policy,
+            &self.retry_params,
         )
         .await?;
         ctx.send_self_message(Loop).await?;
@@ -184,7 +183,8 @@ impl Handler<Loop> for ShardConsumer {
         ctx: &ActorContext<Self>,
     ) -> Result<(), ActorExitStatus> {
         if let Some(shard_iterator) = self.state.next_shard_iterator.take() {
-            let response = get_records(&self.kinesis_client, shard_iterator, &self.retry_policy).await?;
+            let response =
+                get_records(&self.kinesis_client, shard_iterator, &self.retry_params).await?;
             self.state.lag_millis = response.millis_behind_latest.clone();
             self.state.next_shard_iterator = response.next_shard_iterator;
