+++ quickwit-storage/Cargo.toml
@@ -12,6 +12,7 @@ documentation = "https://quickwit.io/docs/"
 [dependencies]
 quickwit-common = { version = "0.2.1", path = "../quickwit-common" }
 quickwit-config = { version = "0.2.1", path = "../quickwit-config" }
+quickwit-aws = {version = "0.2.1", path = "../quickwit-aws"}
 async-trait = '0.1'
 md5 = '0.7'
 anyhow = '1'
