+++ quickwit-storage/src/object_storage/s3_compatible_storage.rs
@@ -29,6 +29,8 @@ use async_trait::async_trait;
 use ec2_instance_metadata::InstanceMetadataClient;
 use futures::{stream, StreamExt};
 use once_cell::sync::OnceCell;
+use quickwit_aws::error::RusotoErrorWrapper;
+use quickwit_aws::retry::{retry, IsRetryable, Retry, RetryParams};
 use quickwit_common::{chunk_range, into_u64_range};
 use regex::Regex;
 use rusoto_core::credential::{AutoRefreshingProvider, ChainProvider, ProfileProvider};
@@ -43,9 +45,7 @@ use tokio::fs::File;
 use tokio::io::{AsyncRead, AsyncReadExt, AsyncWriteExt, BufReader};
 use tracing::{debug, error, info, instrument, warn};
 
-use super::error::RusotoErrorWrapper;
 use crate::object_storage::MultiPartPolicy;
-use crate::retry::{retry, IsRetryable, Retry};
 use crate::{OwnedBytes, Storage, StorageError, StorageErrorKind, StorageResult};
 
 /// A credential timeout.
@@ -188,6 +188,7 @@ pub struct S3CompatibleObjectStorage {
     bucket: String,
     prefix: PathBuf,
     multipart_policy: MultiPartPolicy,
+    retry_params: RetryParams,
 }
 
 impl fmt::Debug for S3CompatibleObjectStorage {
@@ -227,6 +228,7 @@ impl S3CompatibleObjectStorage {
             bucket: bucket.to_string(),
             prefix: PathBuf::new(),
             multipart_policy: MultiPartPolicy::default(),
+            retry_params: RetryParams::default(),
         })
     }
 
@@ -260,6 +262,7 @@ impl S3CompatibleObjectStorage {
             bucket: self.bucket,
             prefix: prefix.to_path_buf(),
             multipart_policy: self.multipart_policy,
+            retry_params: RetryParams::default(),
         }
     }
 
@@ -355,7 +358,10 @@ impl S3CompatibleObjectStorage {
         payload: Box<dyn crate::PutPayload>,
         len: u64,
     ) -> StorageResult<()> {
-        retry(|| self.put_single_part_single_try(key, payload.clone(), len)).await?;
+        retry(&self.retry_params, || {
+            self.put_single_part_single_try(key, payload.clone(), len)
+        })
+        .await?;
         Ok(())
     }
 
@@ -368,7 +374,7 @@ impl S3CompatibleObjectStorage {
             key: key.to_string(),
             ..Default::default()
         };
-        let upload_id = retry(|| async {
+        let upload_id = retry(&self.retry_params, || async {
             self.s3_client
                 .create_multipart_upload(create_upload_req.clone())
                 .await
@@ -472,7 +478,7 @@ impl S3CompatibleObjectStorage {
             stream::iter(parts.into_iter().map(|part| {
                 let payload = payload.clone();
                 let upload_id = upload_id.clone();
-                retry(move || {
+                retry(&self.retry_params, move || {
                     self.upload_part(upload_id.clone(), key, part.clone(), payload.clone())
                 })
             }))
@@ -518,7 +524,7 @@ impl S3CompatibleObjectStorage {
             upload_id: upload_id.to_string(),
             ..Default::default()
         };
-        retry(|| async {
+        retry(&self.retry_params, || async {
             self.s3_client
                 .complete_multipart_upload(complete_upload_req.clone())
                 .await
@@ -535,7 +541,7 @@ impl S3CompatibleObjectStorage {
             upload_id: upload_id.to_string(),
             ..Default::default()
         };
-        retry(|| async {
+        retry(&self.retry_params, || async {
             self.s3_client
                 .abort_multipart_upload(abort_upload_req.clone())
                 .await
@@ -567,7 +573,7 @@ impl S3CompatibleObjectStorage {
     ) -> StorageResult<Vec<u8>> {
         let cap = range_opt.as_ref().map(Range::len).unwrap_or(0);
         let get_object_req = self.create_get_object_request(path, range_opt);
-        let get_object_output = retry(|| async {
+        let get_object_output = retry(&self.retry_params, || async {
             self.s3_client
                 .get_object(get_object_req.clone())
                 .await
@@ -625,7 +631,7 @@ impl Storage for S3CompatibleObjectStorage {
     // TODO implement multipart
     async fn copy_to_file(&self, path: &Path, output_path: &Path) -> StorageResult<()> {
         let get_object_req = self.create_get_object_request(path, None);
-        let get_object_output = retry(|| async {
+        let get_object_output = retry(&self.retry_params, || async {
             self.s3_client
                 .get_object(get_object_req.clone())
                 .await
@@ -649,7 +655,7 @@ impl Storage for S3CompatibleObjectStorage {
             key,
             ..Default::default()
         };
-        retry(|| async {
+        retry(&self.retry_params, || async {
             self.s3_client
                 .delete_object(delete_object_req.clone())
                 .await
@@ -686,7 +692,7 @@ impl Storage for S3CompatibleObjectStorage {
             key,
             ..Default::default()
         };
-        let head_object_output_res = retry(|| async {
+        let head_object_output_res = retry(&self.retry_params, || async {
             self.s3_client
                 .head_object(head_object_req.clone())
                 .await
